requirements.txt
streamlit
langchain>=0.2.17
langchain-classic
langchain-chroma>=0.1.2
langchain-community>=0.2.10
langchain-core>=0.2.20
langchain-experimental>=0.0.62
langchain-groq>=0.1.6
langchain-huggingface>=0.0.3
langchain-ollama>=0.1.0
langchain-text-splitters>=0.2.2
langcodes==3.5.0
langdetect==1.0.9
langsmith
chromadb
pymupdf
python-docx
arabic-reshaper
python-bidi
python-dotenv
sentence-transformers
huggingface-hub
pdfplumber